move compiler tests and parser tests to `/tests/syntax`
