+++ tests/syntax/test_parser.py
@@ -1,8 +1,15 @@
 import ast
 import pytest
+from typing import Any
 
-from .helper import parse_string_helper
 import scenic.syntax.ast as s
+from scenic.syntax.parser import parse_string
+
+
+def parse_string_helper(source: str) -> Any:
+    "Parse string and return Scenic AST"
+    return parse_string(source, "exec")
+
 
 class TestEgoAssign:
     def test_basic(self):
@@ -13,7 +20,7 @@ class TestEgoAssign:
                 assert v == 10
             case _:
                 assert False
-    
+
     def test_with_new(self):
         mod = parse_string_helper("ego = new Object")
         stmt = mod.body[0]
@@ -27,3 +34,16 @@ class TestEgoAssign:
         """Ego is a hard keyword and cannot be used except for ego assignment"""
         with pytest.raises(SyntaxError):
             parse_string_helper("ego, x = 10, 20")
+
+
+class TestNew:
+    def test_basic(self):
+        mod = parse_string_helper("new Object")
+        stmt = mod.body[0]
+        match stmt:
+            case ast.Expr(value=s.New(className=c)):
+                assert c == "Object"
+            case _:
+                assert False
+
+    # TODO(shun): Add tests for specifiers
